// src/admin/enums/admin-user.enums.ts
export enum EmailVerificationStatus {
  VERIFIED = 'VERIFIED',
  NOT_VERIFIED = 'NOT_VERIFIED',
}

export enum ProfileCompletionStatus {
  COMPLETE = 'COMPLETE',
  INCOMPLETE = 'INCOMPLETE',
}

export enum EmailVerificationFilter {
  VERIFIED = 'VERIFIED',
  NOT_VERIFIED = 'NOT_VERIFIED',
}

export enum ProfileCompletionFilter {
  COMPLETE = 'COMPLETE',
  INCOMPLETE = 'INCOMPLETE',
}